Update transfer table filters
